Arbitrary sized images can now be handled
The Vespera II image size is twice as large, in each dimension, compared the Seestar S50 camera.

The Vespera II images are too large for the FFT2D routines. So we do it ourselves, row by row, then column by column. But we parallelize these operations over quarter panels.

The Seestar also needs an image rotation by an extra 180 deg, compared to the Vespara II, for the canonical image view. (N up, E left).

The Seestar provides a plate solution, but it appears to be 180 deg out of whack.

The Vespara II does not provide any plate solutions. So we have to submit to Astrometrics.net, and get back the augmented FITS file.

That augmented FITS file is the basis for saying that the Seestar is 180 deg out of whack - assumes Astrometrics.net is ground truth.