Canonical view - more difficult than you might think.
The Vespera II really stressed the algorithms, and showed that we cannot rely on plate solutions to help us. Geometric distortions of the image affect the transform to WCS too much to allow us to borrow them for our needs.

We are just trying to form a rotated FITS image, such that a central horizontal line has constant Declination, and that Declinations grow from bottom toward image top, and Right Ascension grows from right to left.